Rename files and test class
